Decompose and de-hardcode the loading screen (#45)
Break routes/loading/+page.svelte (~680 lines) into focused subcomponents
and a reactive view-model, and migrate its hardcoded color literals to the
theme CSS vars / color-mix blends.

- Extract LoadingView (loading-view.svelte.ts) owning the boot-sequence state
  machine and the load/retry/enter actions; the reference-data sets become a
  single DRY table. +page.svelte is now a thin shell.
- Split the markup into DiamondMark, LoadingHeader, ProgressBar, ManifestWindow
  (+ ManifestRow), ErrorBanner and EnterButton. ManifestWindow keeps its own
  wheel/RAF/opacity DOM state.
- Migrate ~21 color literals to color-mix over base tokens / semantic vars;
  the dark-on-accent button text now uses --text-on-accent.
- Remove dead LoadingSection.svelte (no longer referenced).
- Add LoadingView unit tests; preserve all data-testids.
- Document the screen in docs/frontend.md.
