+++ docs/frontend.md
@@ -105,3 +105,14 @@ Key decisions and rationale:
 - **Optimistic local apply.** The combat-log writer (`log.ts`) filters by `playerManager.logPreferences`, so saving pushes the draft onto the player manager immediately — the change takes effect for the session regardless of the network result. The server endpoint is not yet implemented, so a failed persist surfaces a toast warning (the change still applies locally) rather than blocking the UI. When the backend lands this becomes a normal success path with no UI change.
 - **Live preview reuses the real log row.** The right-hand preview renders a self-contained sample event stream through the actual `LogRow` component, filtered by the _draft_ preferences — so it looks exactly like the real combat log and visibly reflects each toggle before you save.
 - **All colour flows from the root vars.** The semantic combat-log palette lives in the root `--log-*` CSS variables (`+layout.svelte`); `logColors` in `log-kind.ts` is just the `var(--log-*)` reference map shared by the real log, the preview and the per-type rows. Every neutral/alpha shade in the screen is `color-mix(in srgb, var(--…) N%, transparent)` over a base token (off-white `--text-primary`, `--white`, `--black`, the `--eyebrow` accent) or an existing semantic var (`--text-secondary/-tertiary/-muted`, `--border-subtle/-light/-medium`) — no hard-coded `rgba(…)`, so a theme can restyle the whole screen by overriding the variables alone. Follow-up: the other game screens (and the rest of `log-panel`) still carry pre-existing hard-coded neutrals that should migrate to this same pattern.
+
+## Loading screen (the boot manifest)
+
+The loading screen (`src/routes/loading`) runs the boot sequence that fetches the static reference data the game needs and surfaces it as a sequential "manifest" the player can watch tick through. Like the Options/Challenge screens, it is decomposed into small components per the component-size guideline rather than living as one large `+page.svelte`.
+
+Key decisions and rationale:
+
+- **Load orchestration is a reactive view-model.** `LoadingView` (`loading-view.svelte.ts`) owns the manifest state machine (`phase`, `items`, `activeIndex`, the `completed`/`progressPct`/`title` deriveds) and the `start`/`retryFailed`/`enterGame` actions — mirroring `OptionsView`/`InventoryView`. `+page.svelte` just constructs it and renders the presentational subcomponents (`DiamondMark`, `LoadingHeader`, `ProgressBar`, `ManifestWindow` + `ManifestRow`, `ErrorBanner`, `EnterButton`). This makes the load logic unit-testable without rendering.
+- **The reference-data sets are a single table.** Each set is one row in the `REFERENCE_DATA` array (`key` + `label` + a `loaded()` cache check + a typed `load()`), so adding/removing a set is a one-line edit rather than a copy-pasted block. Concurrent fetches of the same key are de-duplicated through a process-wide in-flight cache so a re-mount mid-load can't double-fetch.
+- **The manifest window keeps its own DOM state.** The centred sliding window — the wheel handler, the `requestAnimationFrame` easing toward the active row, and the per-row opacity falloff — lives inside `ManifestWindow`, driven only by the `items`/`activeIndex`/`phase` props, keeping that view concern out of the load model.
+- **All colour flows from the root vars.** Every neutral/alpha shade is `color-mix(in srgb, var(--…) N%, transparent)` over a base token (`--accent`, `--text-primary`, `--error`, `--success`) or an existing semantic var (`--text-muted`), and the dark-on-accent button text uses `--text-on-accent` — no hard-coded `rgba(…)`/hex, so a theme can restyle the screen by overriding the variables alone.
